Use same store for all components on the page
