+++ app/javascript/packs/actions/index.js
@@ -1,3 +1,5 @@
 import actionTypes from 'constants'
 
-export const sampleSubmit = (name) => ({ type: actionTypes.SAMPLE_ACTION, payload: name });
+export const sampleSubmit = (name) => (
+  { type: actionTypes.CHANGE_NAME, payload: name }
+)
