+++ app/javascript/packs/constants/index.js
@@ -1,7 +1,7 @@
 import mirrorCreator from 'mirror-creator'
 
 const actionTypes = mirrorCreator([
-  'SAMPLE_ACTION',
+  'CHANGE_NAME',
 ])
 
 export default actionTypes
