+++ app/javascript/packs/styles/hello-world.module.scss
